Task 1, 2, 3, 4, 6, 10, 12(improved), 13(improved), 14

+++ src/main/resources/sql/expenses_ddl.sql
@@ -1,7 +1,7 @@
 create table if not exists expenses
 (
 	id bigint primary key auto_increment,
-    expense_sum bigint not null,
+    expense_sum DOUBLE not null,
     expense_date date not null,
     comment varchar(255),
     category_id bigint not null,
