+++ src/main/resources/sql/incomes_ddl.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS incomes
 (
 	id BIGINT PRIMARY KEY AUTO_INCREMENT,
-    income_sum BIGINT NOT NULL,
+    income_sum DOUBLE NOT NULL,
     income_date DATE NOT NULL,
     comment VARCHAR(255)
 );
